RTable: remove observer

+++ RTable/RTable/Categories/NSString+IDPExtensions.m
@@ -10,7 +10,7 @@ static NSString * const kRTRandomStringCharacterSet =
 	NSUInteger characterSetLength = [kRTRandomStringCharacterSet length];
 	
 	for (int i = 0; i < length; ++i) {
-		NSUInteger randomCharacterIndex = arc4random_uniform(characterSetLength);
+		NSUInteger randomCharacterIndex = arc4random_uniform((uint32_t)characterSetLength);
 		unichar randomLetter = [kRTRandomStringCharacterSet characterAtIndex: randomCharacterIndex];
 		[randomString appendFormat: @"%C", randomLetter];
 	}
